Add email functionality, handicap service, and multiple league enhancements
Major features:
- Implement email service infrastructure with Firebase and Mailgun support
- Add handicap calculation service using simplified 4-of-6 method
- Add Firebase upload toggle and duplicate date controls in admin screens
- Implement tied mode for closest pin calculations
- Add league amounts persistence service

UI/UX improvements:
- Restore Skat Purse and Mulligan Purse before autofill recalculation
- Update responsive typography and font size documentation
- Add comprehensive user guides for both Monday and Wednesday leagues

Bug fixes:
- Fix autofill logic to properly reset purses before recalculation
- Update Android manifest for email app queries
- Clean up unused imports and legacy code

Documentation:
- Add email setup guides for Firebase and Mailgun
- Add automatic email implementation summary
- Update CLAUDE.md with new handicap calculation method and league instructions

+++ lib/services/screen_data_retention_service.dart
@@ -28,7 +28,7 @@ class ScreenDataRetentionService {
   bool? _playersShuffled;
   
   // Monday Closest Pin Screen data
-  Map<String, int>? _playerClosestPinCounts;
+  Map<String, double>? _playerClosestPinCounts;
   Map<String, double>? _playerClosestPinWinnings;
   double? _remainingClosestPinPurse;
   int? _totalClosestPins;
@@ -90,7 +90,7 @@ class ScreenDataRetentionService {
 
   /// Captures data from Monday Closest Pin Screen when "Save Results" button is clicked
   void captureClosestPinData({
-    required Map<String, int> playerClosestPinCounts,
+    required Map<String, double> playerClosestPinCounts,
     required Map<String, double> playerClosestPinWinnings,
     required double remainingClosestPinPurse,
     required int totalClosestPins,
@@ -153,7 +153,7 @@ class ScreenDataRetentionService {
   bool? get playersShuffled => _playersShuffled;
   
   /// Gets the player closest pin counts from Monday Closest Pin Screen
-  Map<String, int>? get playerClosestPinCounts => _playerClosestPinCounts;
+  Map<String, double>? get playerClosestPinCounts => _playerClosestPinCounts;
   
   /// Gets the player closest pin winnings from Monday Closest Pin Screen
   Map<String, double>? get playerClosestPinWinnings => _playerClosestPinWinnings;
